requirements.txt
Django==2.2.9
pytz==2019.3
sqlparse==0.3.0
gunicorn==19.9.0
whitenoise==4.1.2
django-crispy-forms==1.8.1

